Corrected frame count for multiple frames per point processing in reruns.
Only one image per point was being processed but frame count was set to images per point -1

Changed to use all the images at each point and correctly set the frame count to number of images per point.

Final analysis fails because there is an explicit test to fail if there are more than 1 images in the final folder.  But there are multiple images in the final folder if there are multiple images per point.  Maybe it should analyse the last of the final images?